+++ src/animator/Animator.ts
@@ -4,33 +4,39 @@ import {Integer} from '../types/Integer'
 import {Milliseconds} from '../types/Milliseconds'
 import {NumberUtil} from '../utils/NumberUtil'
 
-/** Record and update an Atlas.Animation's state. */
+/** Record and update an `Animation`'s state. */
 export interface Animator {
-  /** Cel index oscillation state. This integer may fall outside of animation
-      bounds depending on the animation interval selected by direction. This
-      value should be carried over from each call unless the cel is manually
-      set. Any integer in [0, length - 1] is always valid. */
+  /**
+   * `Cel` index oscillation state. This integer may fall outside of animation
+   * bounds depending on the animation interval selected by direction. This
+   * value should be carried over from each call unless the `Cel` is manually
+   * set. Any integer in [0, `Animation.cels.length`) is always valid.
+   *
+   * Every `Animation` is expected to have at least one `Cel`.
+   */
   period: Integer
 
-  /** Current cel exposure in milliseconds. When the fractional value meets or
-      exceeds the cel exposure duration, the cel is advanced according to
-      direction. This value should be carried over from each call with the
-      current time step added, and zeroed on manual cel change. */
+  /**
+   * Current `Cel` exposure in milliseconds. When the fractional value meets or
+   * exceeds the `Cel` exposure duration, the `Cel` is advanced according to
+   * direction. This value should be carried over from each call with the
+   * current time step added, and zeroed on manual `Cel` change.
+   */
   exposure: Milliseconds
 }
 
 export namespace Animator {
-  /** Apply the time since last frame was shown, possibly advancing the
-      animation period. */
+  /**
+   * Apply the time since last frame was shown, possibly advancing the animation
+   * period.
+   */
   export function animate(
     period: Integer,
     exposure: Milliseconds,
     animation: Atlas.Animation
   ): Animator {
-    if (animation.cels.length < 1) return {period, exposure}
-
     // Avoid unnecessary iterations by skipping complete cycles.
-    // animation.duration may be infinite but the modulo of any number and
+    // `animation.duration` may be infinite but the modulo of any number and
     // infinity is that number.
     exposure = exposure % animation.duration
     while (