+++ changelog.md
@@ -4,6 +4,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ## v3.0.1 (unreleased)
 
+- \[Animator\] Assume all Animations have at least one Cel
 - \[build\] Rebuild the demo assets with Aseprite v1.2.25
 
 ## v3.0.0